rework auto downloader
