+++ NerdNewsNavigator2/Platforms/Android/MainActivity.cs
@@ -11,7 +11,7 @@ namespace NerdNewsNavigator2.Platforms.Android;
 [Activity(Theme = "@style/Maui.SplashTheme", MainLauncher = true, LaunchMode = LaunchMode.SingleInstance, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation | ConfigChanges.UiMode | ConfigChanges.ScreenLayout | ConfigChanges.SmallestScreenSize | ConfigChanges.Density)]
 public class MainActivity : MauiAppCompatActivity
 {
-    private AutoStartService AutoService { get; set; } = new();
+    private AutoDownloadService AutoDownloadService { get; set; } = new();
     public MainActivity()
     {
         var messenger = MauiApplication.Current.Services.GetService<IMessenger>();
@@ -43,12 +43,12 @@ public class MainActivity : MauiAppCompatActivity
     }
     private void StopService()
     {
-        if (AutoService.CancellationTokenSource is not null)
+        if (AutoDownloadService.CancellationTokenSource is not null)
         {
-            AutoService.CancellationTokenSource.Cancel();
-            AutoService.LongTask(AutoService.CancellationTokenSource.Token);
-            AutoService.CancellationTokenSource?.Dispose();
-            AutoService.CancellationTokenSource = null;
+            AutoDownloadService.CancellationTokenSource.Cancel();
+            AutoDownloadService.LongTask(AutoDownloadService.CancellationTokenSource.Token);
+            AutoDownloadService.CancellationTokenSource?.Dispose();
+            AutoDownloadService.CancellationTokenSource = null;
         }
 
         System.Diagnostics.Debug.WriteLine("Stopping AutoDownload");
