+++ NerdNewsNavigator2/Platforms/Windows/App.xaml.cs
@@ -14,11 +14,7 @@ namespace NerdNewsNavigator2.WinUI;
 /// </summary>
 public partial class App : MauiWinUIApplication
 {
-    private readonly System.Timers.Timer _aTimer = new(60 * 60 * 1000);
-    private string WifiOnlyDownloading { get; set; }
-    private string Status { get; set; }
-    public static CancellationTokenSource CancellationTokenSource { get; set; } = null;
-
+    AutoDownloadService AutoDownloadService { get; set; } = new();
     /// <summary>
     /// Initializes the singleton application object.  This is the first line of authored code
     /// executed, and as such is the logical equivalent of main() or WinMain().
@@ -26,7 +22,6 @@ public partial class App : MauiWinUIApplication
     public App()
     {
         InitializeComponent();
-        WifiOnlyDownloading = Preferences.Default.Get("WifiOnly", "No");
         Microsoft.Maui.Handlers.WindowHandler.Mapper.AppendToMapping(nameof(IWindow), (handler, view) =>
         {
             var nativeWindow = handler.PlatformView;
@@ -55,99 +50,11 @@ public partial class App : MauiWinUIApplication
         {
             if (message.Start)
             {
-                Start();
-                Task.Run(() =>
-                {
-                    Thread.Sleep(60 * 1000);
-                    _ = DownloadService.AutoDownload();
-                });
+                AutoDownloadService.Start();
             }
             else
             {
-                Stop();
-            }
-        });
-    }
-
-    /// <summary>
-    /// A method that Auto starts Downloads
-    /// </summary>
-    public void Start()
-    {
-        if (CancellationTokenSource is null)
-        {
-            var cts = new CancellationTokenSource();
-            CancellationTokenSource = cts;
-        }
-        else if (CancellationTokenSource is not null)
-        {
-            CancellationTokenSource.Dispose();
-            CancellationTokenSource = null;
-            var cts = new CancellationTokenSource();
-            CancellationTokenSource = cts;
-        }
-        Debug.WriteLine("Start Auto downloads");
-        LongTask(CancellationTokenSource.Token);
-    }
-    /// <summary>
-    /// A method that Stops auto downloads
-    /// </summary>
-    public void Stop()
-    {
-        if (CancellationTokenSource is not null)
-        {
-            CancellationTokenSource.Cancel();
-            LongTask(CancellationTokenSource.Token);
-            CancellationTokenSource?.Dispose();
-            CancellationTokenSource = null;
-        }
-        Debug.WriteLine("Stopped Auto Downloder");
-    }
-
-    public void LongTask(CancellationToken cancellationToken)
-    {
-        if (cancellationToken.IsCancellationRequested)
-        {
-            _aTimer.Stop();
-            Connectivity.Current.ConnectivityChanged -= GetCurrentConnectivity;
-            _aTimer.Elapsed -= new System.Timers.ElapsedEventHandler(OnTimedEvent);
-            return;
-        }
-        _aTimer.Elapsed += new System.Timers.ElapsedEventHandler(OnTimedEvent);
-        Connectivity.Current.ConnectivityChanged += GetCurrentConnectivity;
-        _aTimer.Start();
-    }
-    private void GetCurrentConnectivity(object sender, ConnectivityChangedEventArgs e)
-    {
-        System.Diagnostics.Debug.WriteLine("Connection status has changed");
-        Status = string.Join(", ", Connectivity.Current.ConnectionProfiles);
-        WifiOnlyDownloading = Preferences.Default.Get("WifiOnly", "No");
-        if (Status == string.Empty)
-        {
-            Debug.WriteLine("No Connection to internet available");
-            return;
-        }
-        System.Diagnostics.Debug.WriteLine(Status);
-    }
-    private void OnTimedEvent(object source, System.Timers.ElapsedEventArgs e)
-    {
-        var connectionStatus = Connectivity.Current.ConnectionProfiles;
-        connectionStatus.ToList().ForEach(item =>
-        {
-            switch (item)
-            {
-                case ConnectionProfile.WiFi:
-                case ConnectionProfile.Ethernet:
-                    Debug.WriteLine($"Timed event: {e} Started");
-                    _ = DownloadService.AutoDownload();
-                    break;
-                case ConnectionProfile.Cellular:
-                    if (WifiOnlyDownloading == "No")
-                    {
-                        Debug.WriteLine($"Timed event: {e} Started");
-                        _ = DownloadService.AutoDownload();
-                    }
-                    break;
+                AutoDownloadService.Stop();
             }
         });
     }
